requirements.txt
aiogram==3.15.0
yt-dlp==2024.11.18
aiofiles==24.1.0
python-dotenv==1.0.1
aiohttp==3.10.11
